requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
attrs==23.2.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
clarabel==0.7.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cvxpy==1.4.2
cycler==0.12.1
distlib==0.3.9
ecdsa==0.19.0
ecos==2.0.13
et_xmlfile==2.0.0
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.55.2
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
gradio==5.9.0
gradio_client==1.5.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
imageio==2.36.1
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaggle==1.5.16
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
narwhals==1.17.0
networkx==3.3
npm==0.1.1
numpy==1.26.2
opencv-python==4.8.1.78
openpyxl==3.1.5
optional-django==0.1.0
orjson==3.10.12
osqp==0.6.5
packaging==24.2
pandas==2.1.3
Pillow==10.1.0
pillow-heif==0.13.1
platformdirs==4.3.6
protobuf==5.29.1
pyarrow==18.1.0
pyasn1==0.6.1
pybind11==2.12.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pyproject-toml==0.0.10
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.19
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
qdldl==0.1.7.post1
referencing==0.34.0
requests==2.32.3
rich==13.9.4
rpds-py==0.18.0
rsa==4.9
ruff==0.8.3
safehttpx==0.1.6
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.11.4
scs==3.2.4.post1
semantic-version==2.10.0
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
streamlit==1.41.0
sympy==1.12
tbb==2021.12.0
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2024.12.12
toml==0.10.2
tomlkit==0.13.2
torch==2.3.0
tornado==6.4.2
tqdm==4.66.1
ttkbootstrap==1.10.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.24.0
virtualenv==20.27.1
watchdog==6.0.0
webencodings==0.5.1
websockets==14.1
wheel==0.45.1
